chore(groupDetails.js): rename groupDetails.js to match import
