+++ client/reducers/groupDetails.js
