+++ client/rootReducer.js
@@ -4,7 +4,7 @@
 
 import { combineReducers } from 'redux';
 import auth from './reducers/auth';
-import groupDetails from './reducers/groupId';
+import groupDetails from './reducers/groupDetails';
 import message from './reducers/message';
 import groups, { groupMemberCount } from './reducers/groups';
